Changed file names and class names to a more sustainable naming convention
Unity has no problem with two files with the same name, providing they are in different directories. However, class names are globalized and have to match the file name. Meaning that two scripts cannot be called the same thing, even in different folders. So instead, I have changed to "platformer_playerMovement" to allow for a more organized larger project