Added the option for additional dialogues
There is an 'additionalDialogues' variable added in DialogueManager, which allows you to set dialogues that play after the 'startingDialogue' had been played. The additional dialogues are universal though, meaning that no matter which path the user took in the starting dialogue chain, the same additional dialogues will be played. This can be improved upon in future.

Additionally, the DialogueManager will no longer accept input if the 'currentDialogue' is set to 'null'.