'additionalDialogue' functionality is no longer universal (closes #8)
Instead of having the 'additionalDialogue' array be a variable in DialogueManager, there is now an array for the DialogueObject and ResponseObject. This means that additional dialogues differ depending on which dialogue path you take.

Also added a 'Reset' button in the inspector so that the game does not need to  be restarted each time.