Added a shop and tower placement
When the player presses Escape, the pauses and the shop opens. The player can then choose to select a tower to buy (providing they can afford it). The tower then follows the mouse until the player clicks. The player cannot be placed on the enemy path.

Also changed some colors, added some better UI and a new wave.