Fixed projectile visual (closes #12)
Bullet will now constantly update their target, meaning they follow it. You can tell that they do not move in a straight path, but this is better than having them miss completely.

Additionally, the damage is now applied to the enemy when the bullet reaches them, instead of when the bullet is spawned.

A new wave was added as well.