Added health display and enemy display
The player now has health. When an enemy reaches the end of the path, the player loses one health. This value is displayed.

There is also now a UI element that tells the player how many enemies have spawned, and how many are in the wave total.

Small tweaks were made to gameplay.