Fixed tower placement collision issue (closes #11)
The issue was not that the rays weren't detecting the collider it was inside, but that both the left & right rays were detecting the same collider. To fix this, the placement is now marked as invalid if the ray01Index and ray02Index are consecutive, or if they are the same.