Improved path visual (closes #13)
The path visual now accounts for diagonal paths. However, seeing as the sprites are rectangles, it doesn't look too great at the corners. Not sure how I would go about fixing this, but it is a rare case anyways. Figuring this out has improved my understanding of rotation and vector math/trigonometry, so I am fairly happy with leaving it as it is.